# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Accessibility Pipeline",
    "version": "0.1.0"
  },
  "paths": {
    "/courses/": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Courses",
        "operationId": "list_courses_courses__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CourseRead"
                  },
                  "type": "array",
                  "title": "Response List Courses Courses  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Create Course",
        "operationId": "create_course_courses__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_id}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Course",
        "operationId": "get_course_courses__course_id__get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "courses"
        ],
        "summary": "Update Course",
        "operationId": "update_course_courses__course_id__patch",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "Delete Course",
        "operationId": "delete_course_courses__course_id__delete",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_id}/artifacts": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Course Artifacts",
        "operationId": "list_course_artifacts_courses__course_id__artifacts_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactRead"
                  },
                  "title": "Response List Course Artifacts Courses  Course Id  Artifacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_id}/review-queue": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Review Queue",
        "operationId": "get_review_queue_courses__course_id__review_queue_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ElementRead"
                  },
                  "title": "Response Get Review Queue Courses  Course Id  Review Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/artifacts/{artifact_id}/elements": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Get Artifact Elements",
        "operationId": "get_artifact_elements_artifacts__artifact_id__elements_get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ElementRead"
                  },
                  "title": "Response Get Artifact Elements Artifacts  Artifact Id  Elements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artifacts/{artifact_id}": {
      "delete": {
        "tags": [
          "artifacts"
        ],
        "summary": "Delete Artifact",
        "operationId": "delete_artifact_artifacts__artifact_id__delete",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/elements/{element_id}": {
      "delete": {
        "tags": [
          "elements"
        ],
        "summary": "Delete Element",
        "operationId": "delete_element_elements__element_id__delete",
        "parameters": [
          {
            "name": "element_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Element Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/element-versions/{version_id}": {
      "get": {
        "tags": [
          "element-versions"
        ],
        "summary": "Get Element Version",
        "operationId": "get_element_version_element_versions__version_id__get",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElementVersionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/element-versions/{version_id}/reviews": {
      "post": {
        "tags": [
          "element-versions"
        ],
        "summary": "Submit Review",
        "operationId": "submit_review_element_versions__version_id__reviews_post",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtifactRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "filename",
          "status",
          "created_at"
        ],
        "title": "ArtifactRead"
      },
...
      "CourseCreate": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 50,
            "title": "Code",
            "examples": [
              "ME 274"
            ]
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "title": "Name",
            "examples": [
              "Dynamics"
            ]
          },
          "term": {
            "type": "string",
            "maxLength": 50,
            "title": "Term",
            "examples": [
              "Spring 2026"
            ]
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "term"
        ],
        "title": "CourseCreate"
      },
      "CourseRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "term": {
            "type": "string",
            "title": "Term"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "term",
          "owner_id",
          "created_at"
        ],
        "title": "CourseRead"
      },
      "CourseUpdate": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "term": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Term"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          }
        },
        "type": "object",
        "title": "CourseUpdate"
      },
      "ElementRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "artifact_id": {
            "type": "string",
            "format": "uuid",
            "title": "Artifact Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "latest_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Version Id"
          },
          "latest_version_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Version Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "artifact_id",
          "kind",
          "sequence",
          "created_at"
        ],
        "title": "ElementRead"
      },
...
      "ElementVersionRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "element_id": {
            "type": "string",
            "format": "uuid",
            "title": "Element Id"
          },
          "version_number": {
            "type": "integer",
            "title": "Version Number"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "content": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "tool_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Used"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewRead"
            },
            "type": "array",
            "title": "Reviews",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "element_id",
          "version_number",
          "status",
          "content",
          "tool_used",
          "confidence_score",
          "rationale",
          "created_at"
        ],
        "title": "ElementVersionRead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewCreate": {
        "properties": {
          "reviewer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Reviewer Id"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "reviewer_id",
          "decision"
        ],
        "title": "ReviewCreate"
      },
      "ReviewRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "reviewer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Reviewer Id"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reviewer_id",
          "decision",
          "comment",
          "created_at"
        ],
        "title": "ReviewRead"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}